Fixed Lazy Reference and Other Issues
