+++ accounts/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import UserProfile
+from .models import UserProfile, CampusAmbassador
 
 
 class UserProfileAdmin(admin.ModelAdmin):
@@ -8,3 +8,10 @@ class UserProfileAdmin(admin.ModelAdmin):
 
 
 admin.site.register(UserProfile, UserProfileAdmin)
+
+class CampusAmbassadorAdmin(admin.ModelAdmin):
+    class Meta:
+        model = CampusAmbassador
+
+
+admin.site.register(CampusAmbassador, CampusAmbassadorAdmin)
