+++ adminportal/urls.py
@@ -3,8 +3,6 @@ from .views import dashboardTeams
 from .views import dashboardUsers, dashboard, downloadExcel, dashboardCas, sendMail
 from django.conf.urls import url
 
-app_name = 'adminportal'
-
 urlpatterns = [
     path('mail', sendMail.as_view(), name='mail'),
     url(r'teams$', dashboardTeams, name='dteams'),
