+++ templates/main/index.html
@@ -36,7 +36,7 @@
 
     </div>
   </div>
-
+  
 </header>
 <!-- About -->
